save log

+++ src/main/java/com/salespage/salespageservice/app/responses/ProductResponse/ProductDataResponse.java
@@ -3,13 +3,15 @@ package com.salespage.salespageservice.app.responses.ProductResponse;
 import com.salespage.salespageservice.domains.entities.Product;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ProductDataResponse {
   private String productId;
 
   private String productName;
 
-  private Long productPrice;
+  private BigDecimal productPrice;
 
   private float totalRate;
 
@@ -22,7 +24,7 @@ public class ProductDataResponse {
   public void assignFromProduct(Product product){
     productId = product.getId().toHexString();
     productName = product.getProductName();
-    productPrice = product.getPrice().longValue();
+    productPrice = product.getPrice();
     sellerUsername = product.getSellerUsername();
     totalRate = product.getRate().getTotalRate();
     avgPoint = product.getRate().getAvgPoint();
